chore: temporarily change app to app-backup
